Correctly fetches user items
